We now paste to the clipboard even if the clipboard is not focused
