+++ client-v2/src/keyboard/index.ts
@@ -1,3 +1,4 @@
+import Raven from 'raven-js';
 import { Store, Dispatch } from 'types/Store';
 import {
   editorCloseAllOverviews,
@@ -9,9 +10,11 @@ import {
 import { ThunkResult } from 'types/Store';
 import Mousetrap from 'mousetrap';
 import { selectFocusState, setFocusState } from 'bundles/focusBundle';
-import paste from './keyboardActionMaps/paste';
+import { RefDrop } from 'util/collectionUtils';
+import { createArticleFragment } from 'shared/actions/ArticleFragments';
 import { moveUp, moveDown } from './keyboardActionMaps/move';
 import { ArticleFragment } from '../shared/types/Collection';
+import { insertClipboardArticleFragment } from 'actions/Clipboard';
 
 type FocusableTypes =
   | 'clipboard'
@@ -63,7 +66,31 @@ export const createKeyboardActionMap = (store: Store): KeyboardBindingMap => ({
   'command+v': {
     title: 'Paste',
     description: 'Paste an entity',
-    action: bindActionMap(store, paste)
+    action: () =>
+      (store.dispatch as Dispatch)(async (dispatch: Dispatch) => {
+        try {
+          if (!navigator || !(navigator as any).clipboard) {
+            throw new Error('No navigator available on paste');
+          }
+          // A temporary any here pending proper typings
+          const content: string = await (navigator as any).clipboard.readText();
+          if (!content) {
+            return;
+          }
+          const contentData: RefDrop = { type: 'REF', data: content };
+          const articleFragment = await dispatch(
+            createArticleFragment(contentData)
+          );
+          if (!articleFragment) {
+            return;
+          }
+          dispatch(
+            insertClipboardArticleFragment('clipboard', 0, articleFragment.uuid)
+          );
+        } catch (e) {
+          Raven.captureMessage(`Paste to clipboard failed: ${e.message}`);
+        }
+      })
   },
   'command+j': {
     title: 'Close all overviews',
